third_party/WebKit/LayoutTests/editing/deleting/delete-character-002-expected.txt

PASS sample.value is "a\n"
PASS sample.selectionStart is 2
PASS sample.selectionEnd is 2
PASS successfullyParsed is true

TEST COMPLETE
